Add support for creating private CTF channels
Fix issue #125

Add util method for parsing args for options. This method takes an array and
searches for any strings in that array beginning with a hyphen. If it
encounters such a string it strips the leading punctuation from that string and
puts it into a new array. After iterating the given array it returns a 2-tuple
containing the new array of options sans leading punctuation and the original
array minus those extracted options.